refactor(mouse): co-locate tests, make dispatch helpers private
Move the button tests into button.rs and the wheel tests into wheel.rs
so each dispatch path's tests live in the same file as the code they
exercise. Shared test infrastructure (CapturedEffects, capture
observers, set_phys_cursor, test_metrics) moves to a cfg(test)
test_support module in mouse.rs, reachable by both child test modules.

With the tests co-located, every dispatch/decide/helper function's only
callers are now in-file, so demote them from pub(super) to private:
dispatch_mouse_buttons, decide_button, effective_drag_cursor,
protocol_mods, to_viewport_point (button) and dispatch_mouse_wheel,
decide_wheel, effects_from_wheel_action, build_wheel_modifiers_horizontal
(wheel). mouse.rs keeps only the shared-helper tests (cell_at_*,
topmost_surface_at).
